feat(refactor): change service entity + incapsulate card component
